demo: use working paths in config and compose
The previously set paths used to work because the Docker image was built with a standard Dockerfile, whilst now it's handled by GoReleaser and ko build, which results in a slightly different setup.

Mainly, the binary is not placedin /app anymore, but in /ko-app, and changing this setting doesn't seem to be possible. Also, the current working directory is not /ko-app, invalidating the use of relative paths.

This commit fixes the issue by using absolute paths everywhere, and also using more standard unix paths like /etc and /var/lib.

Issue: #9